requirements.txt
streamlit
openai
faiss-cpu
numpy
